refactor(125951): :truck: refatora tasks.py para uma pasta
